refactor: split models into separate files
