+++ src/pipelines.rs
@@ -1,6 +1,7 @@
 use crate::client::{AdobeConnector, CloudManagerClient};
 use crate::errors::throw_adobe_api_error;
-use crate::models::{Execution, Pipeline, PipelinesList, PipelinesResponse};
+use crate::models::executions::Execution;
+use crate::models::pipelines::{Pipeline, PipelinesList, PipelinesResponse};
 use crate::HOST_NAME;
 use reqwest::{Error, Method, StatusCode};
 use std::process;
