+++ src/variables.rs
@@ -2,7 +2,8 @@ use crate::client::{AdobeConnector, CloudManagerClient};
 use crate::encryption::decrypt;
 use crate::environments::get_environment;
 use crate::errors::throw_adobe_api_error;
-use crate::models::{Variable, VariableType, VariablesList, VariablesResponse, YamlConfig};
+use crate::models::config::YamlConfig;
+use crate::models::variables::{Variable, VariableType, VariablesList, VariablesResponse};
 use crate::pipelines::get_pipeline;
 use crate::HOST_NAME;
 use colored::*;
